refactor: extract remark from page
